Add files via upload
So this pops out a text box asking you to guess a state in US. If you get the name of the state right then that state is shown in the map. For those which you couldn't remember the name, after exiting this creates a file called states_to _learn where you can find the list of states that you missed. I used pandas module here. This was a follow along project for me however I am trying to build a similar one for guessing states of India too. So stay around and have fun quizzing this one. Thank you so much.